+++ testcases/open_posix_testsuite/conformance/interfaces/sigprocmask/8-2.c
@@ -44,7 +44,7 @@ int is_changed(sigset_t set, int sig)
 	if (sigismember(&set, sig) != 1) {
 		return 1;
 	}
-	for (i = 0; i < NUMSIGNALS; i++) {
+	for (i = 0; i < (int)NUMSIGNALS; i++) {
 		if ((siglist[i] != sig)) {
 			if (sigismember(&set, siglist[i]) != 0) {
 				return 1;
